refactor: comparison, inquiry 폴더링 수정
