+++ src/main/java/FITPET/dev/controller/AdminController.java
@@ -97,76 +97,6 @@ public class AdminController {
         return ApiResponse.SuccessResponse(SuccessStatus.GET_DELETED_INSURANCES, adminService.getDeletedInsurances(page));
     }
 
-
-
-    /*
-     * Comparison
-     */
-    @GetMapping("/comparison")
-    @Operation(summary = "견적 요청 리스트 조회 API", description = "시작 날짜, 마지막 날짜, 상태값을 Parameter으로 받아 특정 기간 동안 생성된 견적 요청 정보를 조회")
-    public ApiResponse getPetInfos(
-            @RequestParam(name = "page", required = false, defaultValue = "0") int page,
-            @RequestParam(name = "startDate", required = false) String startDate,
-            @RequestParam(name = "endDate", required = false) String endDate,
-            @RequestParam(name = "status", required = false, defaultValue = "PENDING") ComparisonStatus status
-    ) {
-        return ApiResponse.SuccessResponse(SuccessStatus.GET_COMPARISON_TABLE, adminService.getComparisons(startDate, endDate, page, status));
-    }
-
-
-    @GetMapping("/comparison/downloads")
-    @Operation(summary = "견적 요청 리스트 엑셀 다운로드 API", description = "시작 날짜, 마지막 날짜, 상태값을 Parameter으로 받아 특정 기간 동안 생성된 견적 요청 정보를 가진 엑셀을 다운로드")
-    public void downloadPetInfos(HttpServletResponse servletResponse,
-                                 @RequestParam(name = "startDate") String startDate,
-                                 @RequestParam(name = "endDate") String endDate,
-                                 @RequestParam(name = "status", required = false, defaultValue = "PENDING") ComparisonStatus status) {
-        adminService.downloadComparisons(servletResponse, startDate, endDate, status);
-    }
-
-
-    @PatchMapping("/comparison/status/{comparisonId}")
-    @Operation(summary = "견적 요청 상태 변경 API", description = "특정 견적 요청 정보의 상태값을 변경")
-    public ApiResponse patchPetInfoStatus(
-            @PathVariable(value = "comparisonId") Long comparisonId,
-            @RequestParam(name = "status") ComparisonStatus status
-    ) {
-        return ApiResponse.SuccessResponse(SuccessStatus.PATCH_COMPARISON_STATUS, adminService.patchComparisonStatus(comparisonId, status));
-    }
-
-
-    @GetMapping("/comparison/search")
-    @Operation(summary = "견적서 검색 API", description = "전화번호와 펫 이름으로 견적서 검색")
-    public ApiResponse searchPetInfos(
-            @RequestParam(name = "content", required = false) String content,
-            @RequestParam(name = "page", required = false, defaultValue = "0") int page
-    ) {
-        return ApiResponse.SuccessResponse(SuccessStatus.SEARCH_COMPARISON, adminService.searchComparisons(content, page));
-    }
-
-    /*
-     * Inquiry
-     */
-    @GetMapping("/inquiry")
-    @Operation(summary = "1:1 문의 내역 조회 API", description = "시작 날짜, 마지막 날짜, 상태값을 Parameter으로 받아 특정 기간 동안 생성된 1:1 문의 내역 조회")
-    public ApiResponse getInquiries(
-            @RequestParam(name = "startDate", required = false) String startDate,
-            @RequestParam(name = "endDate", required = false) String endDate,
-            @RequestParam(name = "status", required = false) InquiryStatus inquiryStatus
-    ){
-        return ApiResponse.SuccessResponse(SuccessStatus.GET_INQUIRY, adminService.getInquiries(startDate, endDate, inquiryStatus));
-    }
-
-
-    @PatchMapping("/inquiry/status/{inquiryId}")
-    @Operation(summary = "1:1 문의 상태 변경 API", description = "특정 1:1 문의 정보의 상태값을 변경")
-    public ApiResponse patchInquiryStatus(
-            @PathVariable(value = "inquiryId") Long inquiryId,
-            @RequestParam(name = "status") InquiryStatus inquiryStatus
-    ) {
-        return ApiResponse.SuccessResponse(SuccessStatus.PATCH_COMPARISON_STATUS, adminService.patchInquiryStatus(inquiryId, inquiryStatus));
-    }
-
-
     /*
      * Proposal
      */
